#1 :recycle: carousel logic -> useCarousel (custom hook)
